backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via
    #   httpcore
    #   starlette
async-timeout==4.0.3
    # via redis
billiard==4.1.0
    # via celery
celery==5.3.1
    # via -r requirements.in
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
click==8.1.6
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
fastapi==0.101.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
iniconfig==2.0.0
    # via pytest
kombu==5.3.1
    # via celery
packaging==23.1
    # via
    #   pytesseract
    #   pytest
pillow==10.0.0
    # via pytesseract
pluggy==1.2.0
    # via pytest
prompt-toolkit==3.0.39
    # via click-repl
pydantic==2.1.1
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.4.0
    # via pydantic
pytesseract==0.3.10
    # via -r requirements.in
pytest==7.4.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via celery
python-multipart==0.0.6
    # via -r requirements.in
redis==4.6.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.27.0
    # via fastapi
typing-extensions==4.7.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
tzdata==2023.3
    # via celery
uvicorn==0.23.2
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
websockets==11.0.3
    # via -r requirements.in
